Split long functions and modules along seams
Two structural moves, no behavior change, and the same check standing
behind both: every test count is what it was.

Inside files, report_prompt separates the brief, which interpolates
every value, from the rules, which interpolate only the rubric version;
the golden fixture in tests/golden/prompts.json proves the text is
unchanged to the byte, and caught two newlines at the seam while it was
being made. apply_control becomes one function per control message, and
run_web hands the configuration it assembles to web_server_config.

Across files, src/livekit.rs was the last module the splits that made
src/recording/ and src/web/ never reached. Two regions its own module
doc already named come out whole with their tests: rooms is the only
part that talks to LiveKit over HTTP rather than over the room's event
stream, and report is the packet an ending interview produces. Session
setup stays, and the module doc now says why, because open_session
reaches for five of the loop's own types and six of its functions.
